[fix] 新規ユーザ作成時にもログインループが発生しないように修正
